+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- hotfix([237](https://github.com/meateam/api-gateway/pull/237)): add pending status for transfer
 - hotfix([236](https://github.com/meateam/api-gateway/pull/236)): fix delete file in shared folder & fix transfer info premission
 
 ## [v4.4.0] - 2021-05-18